feat: Launch page responsive & animated
