+++ next.config.js
@@ -4,7 +4,7 @@ const nextConfig = {
     return config
   },
   env: {
-    CONTRACT_ADDRESS: "0x1d460d731Bd5a0fF2cA07309dAEB8641a7b175A1",
+    CONTRACT_ADDRESS: "0x2f321ed425c82E74925488139e1556f9B76a2551",
     WETH_ADDRESS: "0xc02aaa39b223fe8d0a0e5c4f27ead9083c756cc2",
     UNISWAP_ROUTER_ADDRESS: "0x7a250d5630B4cF539739dF2C5dAcb4c659F2488D",
     API_ENDPOINT: 'http://localhost:8000/subgraphs/name/SAFUlauncher'
