+++ src/static/templates.json
@@ -1,24 +1,146 @@
 {
-    "USDSC":
+    "SAFU":
     {
-      "minLiq": 0,
-      "supply": 10000000,
-      "initTaxType": 0,
-      "initInterval": 20,
-      "countInterval": 2,
-      "maxBuyTax": 20,
-      "minBuyTax": 1,
-      "maxSellTax": 20,
-      "minSellTax": 1,
-      "lpTax": 10,
-      "maxWallet": 1,
-      "maxTx": 1,
-      "preventSwap": 10,
-      "maxSwap": 10000,
-      "taxSwapThreshold": 1000,
-      "cliffPeriod": 30,
-      "vestingPeriod": 30,
-      "name": "USDCoin",
-      "symbol": "USDC"
-    }
+			"stakingFacet": "0x2f321ed425c82E74925488139e1556f9B76a2551",
+			"isFreeTier": true,
+			"minLiq": 0,
+			"supply": 1000000,
+			"initTaxType": 0,
+			"initInterval": 0,
+			"countInterval": 0,
+			"maxBuyTax": 40,
+			"minBuyTax": 5,
+			"maxSellTax": 40,
+			"minSellTax": 5,
+			"lpTax": 0,
+			"maxWallet": 1,
+			"maxTx": 1,
+			"preventSwap": 10,
+			"maxSwap": 10000,
+			"taxSwapThreshold": 1000,
+			"cliffPeriod": 30,
+			"vestingPeriod": 30,
+			"name": "SAFU",
+			"symbol": "SAFU"
+		},
+    "USDC":
+    {
+			"stakingFacet": "0x2f321ed425c82E74925488139e1556f9B76a2551",
+			"isFreeTier": true,
+			"minLiq": 0,
+			"supply": 1000000,
+			"initTaxType": 0,
+			"initInterval": 0,
+			"countInterval": 0,
+			"maxBuyTax": 40,
+			"minBuyTax": 5,
+			"maxSellTax": 40,
+			"minSellTax": 5,
+			"lpTax": 0,
+			"maxWallet": 1,
+			"maxTx": 1,
+			"preventSwap": 10,
+			"maxSwap": 10000,
+			"taxSwapThreshold": 1000,
+			"cliffPeriod": 30,
+			"vestingPeriod": 30,
+			"name": "USD Coin",
+			"symbol": "USDC"
+		},
+    "WBTC":
+    {
+			"stakingFacet": "0x2f321ed425c82E74925488139e1556f9B76a2551",
+			"isFreeTier": true,
+			"minLiq": 0,
+			"supply": 1000000,
+			"initTaxType": 0,
+			"initInterval": 0,
+			"countInterval": 0,
+			"maxBuyTax": 40,
+			"minBuyTax": 5,
+			"maxSellTax": 40,
+			"minSellTax": 5,
+			"lpTax": 0,
+			"maxWallet": 1,
+			"maxTx": 1,
+			"preventSwap": 10,
+			"maxSwap": 10000,
+			"taxSwapThreshold": 1000,
+			"cliffPeriod": 30,
+			"vestingPeriod": 30,
+			"name": "Wrapped BTC",
+			"symbol": "WBTC"
+		},
+    "USDT":
+    {
+			"stakingFacet": "0x2f321ed425c82E74925488139e1556f9B76a2551",
+			"isFreeTier": true,
+			"minLiq": 0,
+			"supply": 1000000,
+			"initTaxType": 0,
+			"initInterval": 0,
+			"countInterval": 0,
+			"maxBuyTax": 40,
+			"minBuyTax": 5,
+			"maxSellTax": 40,
+			"minSellTax": 5,
+			"lpTax": 0,
+			"maxWallet": 1,
+			"maxTx": 1,
+			"preventSwap": 10,
+			"maxSwap": 10000,
+			"taxSwapThreshold": 1000,
+			"cliffPeriod": 30,
+			"vestingPeriod": 30,
+			"name": "Tether USD",
+			"symbol": "USDT"
+		},
+    "MAGA":
+    {
+			"stakingFacet": "0x2f321ed425c82E74925488139e1556f9B76a2551",
+			"isFreeTier": true,
+			"minLiq": 0,
+			"supply": 1000000,
+			"initTaxType": 0,
+			"initInterval": 0,
+			"countInterval": 0,
+			"maxBuyTax": 40,
+			"minBuyTax": 5,
+			"maxSellTax": 40,
+			"minSellTax": 5,
+			"lpTax": 0,
+			"maxWallet": 1,
+			"maxTx": 1,
+			"preventSwap": 10,
+			"maxSwap": 10000,
+			"taxSwapThreshold": 1000,
+			"cliffPeriod": 30,
+			"vestingPeriod": 30,
+			"name": "MAGA",
+			"symbol": "MAGA"
+		},
+    "MOG":
+    {
+			"stakingFacet": "0x2f321ed425c82E74925488139e1556f9B76a2551",
+			"isFreeTier": true,
+			"minLiq": 0,
+			"supply": 1000000,
+			"initTaxType": 0,
+			"initInterval": 0,
+			"countInterval": 0,
+			"maxBuyTax": 40,
+			"minBuyTax": 5,
+			"maxSellTax": 40,
+			"minSellTax": 5,
+			"lpTax": 0,
+			"maxWallet": 1,
+			"maxTx": 1,
+			"preventSwap": 10,
+			"maxSwap": 10000,
+			"taxSwapThreshold": 1000,
+			"cliffPeriod": 30,
+			"vestingPeriod": 30,
+			"name": "Mog Coin",
+			"symbol": "MOG"
+		},
 }
